Merge: messaging: enable HA queues when HAproxy binding is enabled
If HAproxy binding is enabled for RabbitMQ service, we need to ensure
queres are replicated and this is done by a specific policy.

Since OpenStack is already tagging queues with rabbit_ha_queues=True
parameter everywhere, we just create the policy and the queues by
themselves.

This is enabled only if HAproxy binding is enabled for now, because the
feature seems required when using RabbitMQ load-balancing. The other
use-case (when using Oslo messaging to failover to another RabbitMQ node)
does not require this policy. If a node is down, OpenStack will fallback
to another RabbitMQ node in the list from parameters.

This patch allows you to configure 'exactly' or 'all' RabbitMQ HA policy modes.

Why using 'exactly' mode ?
To let operators scale up/down the replication accross an exact number
of nodes.

How it works ?
Queue is mirrored to count nodes in the cluster. If there are less than
count nodes in the cluster, the queue is mirrored to all nodes. If there
are more than count nodes in the cluster, and a node containing a mirror
goes down, then a new mirror will be created on another node.

Also change the Puppetfile to have latest commits in puppetlabs-rabbitmq
to have some fix related to rabbitmq_policy provider.
Without this fix, some parameters are not integers but string, and we
need integer.

Closes-bug #1434474

Change-Id: I3f732360f83e10e0ae525573f9afdb7f68475149